+++ include/snapclient_dsp_store.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "snapclient_dsp.h"
+
+app_config::SnapclientDspConfig loadSnapclientDspConfig();
+void saveSnapclientDspConfig(const app_config::SnapclientDspConfig &config);
+void resetSnapclientDspConfig();